Create basic tab screens and clean project
